requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
identify==2.6.1
kombu==5.4.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
psycopg==3.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
ruff==0.7.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
